extract business logic from AuthController to AuthService and replace magic strings with constants
