+++ SubscriptionManager/SubscriptionManager.Core/Constants/ClaimTypesConstants.cs
