+++ SubscriptionManager/SubscriptionManager.Core/Constants/RoleConstants.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SubscriptionManager.Core.Constants
+{
+    public static class RoleConstants
+    {
+        public const string User = "User";
+        public const string Admin = "Admin";
+        public const string DefaultRole = User;
+    }
+}
