+++ SubscriptionManager/SubscriptionManager.Core/Models/Responses/AuthResult.cs
@@ -10,5 +10,5 @@ public class AuthResult
 {
     public bool Success { get; set; }
     public string? UserId { get; set; }
-    public List<string> Errors { get; set; } = new();
+    public string? Error { get; set; }
 }
