+++ SubscriptionManager/SubscriptionManager.Core/Models/Responses/LoginResponse.cs
@@ -12,5 +12,5 @@ public class LoginResponse
     public string? AccessToken { get; set; }
     public string? RefreshToken { get; set; }
     public DateTime AccessTokenExpiresAt { get; set; }
-    public List<string> Errors { get; set; } = new();
+    public string? Error { get; set; }
 }
