+++ SubscriptionManager/SubscriptionManager.Core/Models/Responses/RefreshTokenResponse.cs
@@ -12,5 +12,5 @@ public class RefreshTokenResponse
     public string? AccessToken { get; set; }
     public string? RefreshToken { get; set; }
     public DateTime AccessTokenExpiresAt { get; set; }
-    public List<string> Errors { get; set; } = new();
+    public string? Error { get; set; }
 }
